refactor: mv files to src dir
