+++ src/authentication/src/steps/Welcome.jsx
